requirements.txt
python-telegram-bot==20.7
yfinance
pandas
numpy
requests